# openapi/core-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/palm-match/reading": {
      "get": {
        "operationId": "PalmMatchController_getReading",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["love", "career", "health", "general"]
            }
          },
          {
            "name": "targetName",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchReadingDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Nội dung bói toán giải trí — deterministic theo user + category + ngày server (guest dùng được)",
        "tags": ["palm-match"]
      }
    },
    "/api/v1/palm-match/queue": {
      "post": {
        "operationId": "PalmMatchController_joinQueue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchStateDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Enqueue idempotent và thử ghép Palm Match ẩn danh",
        "tags": ["palm-match"]
      }
    },
    "/api/v1/palm-match/current": {
      "get": {
        "operationId": "PalmMatchController_getCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchStateDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "State phục hồi/poll của queue hoặc session hiện tại",
        "tags": ["palm-match"]
      },
      "delete": {
        "operationId": "PalmMatchController_dismissCurrent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchStateDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Huỷ queue/session active hoặc dismiss kết quả terminal",
        "tags": ["palm-match"]
      }
    },
    "/api/v1/palm-match/sessions/{id}/flip": {
      "post": {
        "operationId": "PalmMatchController_flip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchStateDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Lật lá của chính caller — idempotent",
        "tags": ["palm-match"]
      }
    },
    "/api/v1/palm-match/sessions/{id}/rating": {
      "post": {
        "operationId": "PalmMatchController_rate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RatePalmMatchDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PalmMatchStateDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Chốt like/skip một lần; mutual-like mở Friendship",
        "tags": ["palm-match"]
      }
    },
...
    "/api/v1/support/tickets": {
      "post": {
        "operationId": "SupportController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/SupportTicketDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Gửi phản hồi/yêu cầu hỗ trợ — idempotent",
        "tags": ["support"]
      }
    },
    "/api/v1/support/tickets/me": {
      "get": {
        "operationId": "SupportController_listMine",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SupportTicketsPageDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Theo dõi các yêu cầu hỗ trợ của chính mình",
        "tags": ["support"]
      }
    },
...
  },
  "info": {
    "title": "Litmatch core-api",
    "description": "",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "PalmMatchReadingDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": ["love", "career", "health", "general"]
          },
          "content": { "type": "string" },
          "forDate": {
            "type": "string",
            "description": "Ngày server (UTC, YYYY-MM-DD) dùng làm seed"
          }
        },
        "required": ["category", "content", "forDate"]
      },
      "PalmZodiacSignDto": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "symbol": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["key", "symbol", "name"]
      },
      "PalmMatchStateDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": ["idle", "queued", "active", "completed"]
          },
          "sessionId": { "type": "string", "format": "uuid" },
          "queuedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "myFlipped": { "type": "boolean" },
          "opponentFlipped": { "type": "boolean" },
          "mySign": { "$ref": "#/components/schemas/PalmZodiacSignDto" },
          "opponentSign": { "$ref": "#/components/schemas/PalmZodiacSignDto" },
          "compatibilityPercent": {
            "type": "number",
            "minimum": 60,
            "maximum": 99
          },
          "fortune": { "type": "string" },
          "myRating": { "type": "string", "enum": ["like", "skip"] },
          "outcome": {
            "type": "string",
            "enum": ["matched", "not_matched", "expired", "cancelled"]
          },
          "partnerUserId": {
            "type": "string",
            "description": "Chỉ xuất hiện sau mutual-like"
          }
        },
        "required": ["state"]
      },
      "RatePalmMatchDto": {
        "type": "object",
        "properties": {
          "rating": { "type": "string", "enum": ["like", "skip"] }
        },
        "required": ["rating"]
      },
...
      "SupportTicketDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "category": { "type": "string", "enum": ["feedback", "bug", "idea"] },
          "message": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["open", "in_progress", "resolved", "closed"]
          },
          "staffResponse": { "type": "string", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "category",
          "message",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateSupportTicketDto": {
        "type": "object",
        "properties": {
          "category": { "type": "string", "enum": ["feedback", "bug", "idea"] },
          "message": { "type": "string", "minLength": 5, "maxLength": 2000 }
        },
        "required": ["category", "message"]
      },
      "SupportTicketsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SupportTicketDto" }
          },
          "meta": {
            "type": "object",
            "properties": {
              "nextCursor": { "type": "string", "nullable": true }
            }
          }
        },
        "required": ["items", "meta"]
      },
...
    }
  }
}